pyproject.toml
[tool.poetry]
name = "data-project"
version = "0.1.1"
description = "Crypto Token small data project"
authors = ["Juan Prina <jpprina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^3.0.0"
click = "^8.1.7"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"
isort = "^5.13.2"
pydocstyle = "^6.3.0"
pre-commit = "^3.7.0"


[tool.black]
target-version = ["py39"]
line-length = 99
color = true

[tool.isort]
py_version = 39
line_length = 99
multi_line_output = 3
indent = 4
color_output = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
